split system.c/h into different components, system.h being root include
