+++ src/base/filesystem.h
@@ -1,0 +1,150 @@
+/* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
+/* If you are missing that file, acquire a complete release at teeworlds.com.                */
+
+/*
+	Title: OS Abstraction - Filesystem
+*/
+
+#ifndef BASE_FILESYSTEM_H
+#define BASE_FILESYSTEM_H
+
+#include "detect.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*
+	Function: fs_listdir
+		Lists the files in a directory
+
+	Parameters:
+		dir - Directory to list
+		cb - Callback function to call for each entry
+		type - Type of the directory
+		user - Pointer to give to the callback
+*/
+typedef int (*FS_LISTDIR_CALLBACK)(const char *name, int is_dir, int dir_type, void *user);
+void fs_listdir(const char *dir, FS_LISTDIR_CALLBACK cb, int type, void *user);
+
+/*
+	Function: fs_makedir
+		Creates a directory
+
+	Parameters:
+		path - Directory to create
+
+	Returns:
+		Returns 0 on success. Negative value on failure.
+
+	Remarks:
+		Does not create several directories if needed. "a/b/c" will result
+		in a failure if b or a does not exist.
+*/
+int fs_makedir(const char *path);
+
+/*
+	Function: fs_makedir_recursive
+		Recursively create directories
+
+	Parameters:
+		path - Path to create
+
+	Returns:
+		Returns 0 on success. Negative value on failure.
+*/
+int fs_makedir_recursive(const char *path);
+
+
+/*
+	Function: fs_storage_path
+		Fetches per user configuration directory.
+
+	Returns:
+		Returns 0 on success. Negative value on failure.
+
+	Remarks:
+		- Returns ~/.appname on UNIX based systems
+		- Returns ~/Library/Applications Support/appname on Mac OS X
+		- Returns %APPDATA%/Appname on Windows based systems
+*/
+int fs_storage_path(const char *appname, char *path, int max);
+
+/*
+	Function: fs_is_dir
+		Checks if directory exists
+
+	Returns:
+		Returns 1 on success, 0 on failure.
+*/
+int fs_is_dir(const char *path);
+
+/*
+	Function: fs_chdir
+		Changes current working directory
+
+	Returns:
+		Returns 0 on success, 1 on failure.
+*/
+int fs_chdir(const char *path);
+
+/*
+	Function: fs_getcwd
+		Gets the current working directory.
+
+	Returns:
+		Returns a pointer to the buffer on success, 0 on failure.
+*/
+char *fs_getcwd(char *buffer, int buffer_size);
+
+/*
+	Function: fs_parent_dir
+		Get the parent directory of a directory
+
+	Parameters:
+		path - The directory string
+
+	Returns:
+		Returns 0 on success, 1 on failure.
+
+	Remarks:
+		- The string is treated as zero-terminated string.
+*/
+int fs_parent_dir(char *path);
+
+/*
+	Function: fs_remove
+		Deletes the file with the specified name.
+
+	Parameters:
+		filename - The file to delete
+
+	Returns:
+		Returns 0 on success, 1 on failure.
+
+	Remarks:
+		- The strings are treated as zero-terminated strings.
+*/
+int fs_remove(const char *filename);
+
+/*
+	Function: fs_rename
+		Renames the file or directory. If the paths differ the file will be moved.
+
+	Parameters:
+		oldname - The actual name
+		newname - The new name
+
+	Returns:
+		Returns 0 on success, 1 on failure.
+
+	Remarks:
+		- The strings are treated as zero-terminated strings.
+*/
+int fs_rename(const char *oldname, const char *newname);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
