+++ src/base/memory.c
@@ -1,0 +1,79 @@
+/* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
+/* If you are missing that file, acquire a complete release at teeworlds.com.                */
+#include <stdlib.h>
+#include <string.h>
+
+#include "debug.h"
+#include "memory.h"
+
+#if defined(__cplusplus)
+extern "C" {
+#endif
+
+typedef struct MEMHEADER
+{
+	const char *filename;
+	int line;
+	int size;
+	struct MEMHEADER *prev;
+	struct MEMHEADER *next;
+} MEMHEADER;
+
+typedef struct MEMTAIL
+{
+	int guard;
+} MEMTAIL;
+
+static struct MEMHEADER *first = 0;
+static const int MEM_GUARD_VAL = 0xbaadc0de;
+
+void *mem_alloc_debug(const char *filename, int line, unsigned size, unsigned alignment)
+{
+	return malloc(size);
+}
+
+void mem_free(void *p)
+{
+	free(p);
+}
+
+void mem_copy(void *dest, const void *source, unsigned size)
+{
+	memcpy(dest, source, size);
+}
+
+void mem_move(void *dest, const void *source, unsigned size)
+{
+	memmove(dest, source, size);
+}
+
+void mem_zero(void *block, unsigned size)
+{
+	memset(block, 0, size);
+}
+
+int mem_comp(const void *a, const void *b, int size)
+{
+	return memcmp(a, b, size);
+}
+
+int mem_check_imp()
+{
+	MEMHEADER *header = first;
+	while(header)
+	{
+		MEMTAIL *tail = (MEMTAIL *)(((char*)(header+1))+header->size);
+		if(tail->guard != MEM_GUARD_VAL)
+		{
+			dbg_msg("mem", "Memory check failed at %s(%d): %d", header->filename, header->line, header->size);
+			return 0;
+		}
+		header = header->next;
+	}
+
+	return 1;
+}
+
+#if defined(__cplusplus)
+}
+#endif
